🚚 move 크라켄 제외 보스 -> 몬스터 폴더로 이동
