+++ Content/_AbyssDiver/Maps/Prototypes_Test/HSRTest/HSR_Test2.umap
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:79fb3111c5fa567677ecac607f14eb9f7d559d46983cf61c25c835540328783b
-size 294584
+oid sha256:2c0c68201441b9878fc7326ec69e96720c370db675cfd78f886ad05114eb5704
+size 285013
