+++ Source/AbyssDiverUnderWorld/Monster/AlienShark/AlienShark.h
@@ -1,7 +1,7 @@
 ﻿#pragma once
 
 #include "CoreMinimal.h"
-#include "Monster/Boss/Boss.h"
+#include "Monster/Monster.h"
 
 #include "AlienShark.generated.h"
