+++ Source/AbyssDiverUnderWorld/Monster/BT/EyeStalker/BTTask_EyeStalkerDetected.cpp
@@ -1,8 +1,8 @@
 #include "Monster/BT/EyeStalker/BTTask_EyeStalkerDetected.h"
 
 #include "Monster/Effect/PostProcessSettingComponent.h"
-#include "Monster/Boss/EyeStalker/EyeStalker.h"
-#include "Monster/Boss/EyeStalker/EyeStalkerAIController.h"
+#include "Monster/EyeStalker/EyeStalker.h"
+#include "Monster/EyeStalker/EyeStalkerAIController.h"
 
 #include "Character/UnderwaterCharacter.h"
 #include "Container/BlackboardKeys.h"
