+++ Source/AbyssDiverUnderWorld/Monster/BT/Limadon/BTTask_LimadonIdle.cpp
@@ -1,7 +1,7 @@
 #include "Monster/BT/Limadon/BTTask_LimadonIdle.h"
 
 //#include "Monster/Boss/Enum/EBossState.h"
-#include "Monster/Boss/Limadon/Limadon.h"
+#include "Monster/Limadon/Limadon.h"
 
 #include "Character/StatComponent.h"
 #include "Character/UnderwaterCharacter.h"
