+++ Source/AbyssDiverUnderWorld/Monster/BT/Serpmare/BTTask_SerpmareIdle.cpp
@@ -3,7 +3,7 @@
 #include "AbyssDiverUnderWorld.h"
 
 //#include "Monster/Boss/Enum/EBossState.h"
-#include "Monster/Boss/Serpmare/Serpmare.h"
+#include "Monster/Serpmare/Serpmare.h"
 #include "Monster/MonsterAIController.h"
 
 UBTTask_SerpmareIdle::UBTTask_SerpmareIdle()
