+++ Source/AbyssDiverUnderWorld/Monster/Limadon/Limadon.cpp
@@ -1,4 +1,4 @@
-#include "Monster/Boss/Limadon/Limadon.h"
+#include "Monster/Limadon/Limadon.h"
 #include "AbyssDiverUnderWorld.h"
 #include "Monster/Boss/Enum/EBossState.h"
 #include "Character/StatComponent.h"
